+++ package/pack.sh
@@ -5,4 +5,4 @@ fvtt package workon "pf2e-kineticists-companion" --type "Module"
 rm -r packs
 
 fvtt package pack effects --in packs-source/effects --out packs
-fvtt package pack items --in packs-source/effects --out items
+fvtt package pack items --in packs-source/items --out packs