+++ scripts/pf2e-kineticists-companion.js
@@ -1,3 +1,4 @@
+import { ArmorInEarth } from "./actions/armor-in-earth.js";
 import { FreshProduce } from "./actions/fresh-produce.js";
 import { ThermalNimbus } from "./actions/thermal-nimbus.js";
 
@@ -10,5 +11,6 @@ Hooks.on(
 
         ThermalNimbus.initialise();
         FreshProduce.initialise();
+        ArmorInEarth.initialise();
     }
 );