+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## Unreleased
+- Add automation for Armor in Earth.
+
 ## 1.1.0 - 2024-10-29 (Pathfinder 2e 6.5.1)
 - Add automation for Fresh Produce.